refactor: move test to keck server
