+++ apps/keck/Cargo.toml
@@ -64,6 +64,7 @@ jwst-storage = { path = "../../libs/jwst-storage" }
 
 [dev-dependencies]
 axum-test-helper = "0.2.0"
+reqwest = { version = "0.11.14", default-features = false, features = ["json", "rustls-tls"] }
 
 [build-dependencies]
 mdbook = { version = "0.4.27", default-features = false, optional = true }
